rename: 대시보드 관련 쿼리 훅 경로 수정
